+++ src/rml_eigen_value_solver_conf.cpp
@@ -9,8 +9,8 @@ static REigenValueSolverDesc eigenValueSolverDesc [] =
 {
     { "None" },
     { "Lanczos" },
-    { "Arnoldi" },
-    { "Rayleigh" }
+    { "Subspace iteration" },
+    { "Inverse power iteration" }
 };